refactor: 기존 패키지에 MVC 패턴 네이밍 적용
