+++ src/main/kotlin/racingcar/model/Cars.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.model
 
 import racingcar.util.RandomGenerator
